perf: C++ core concurrency optimization (atomic & shared_mutex) (#340)
* perf: optimize C++ wrapper with shared_mutex and atomic status flags

This improves concurrency and reduces latency in the C++ wrapper layer by replacing exclusive mutexes with shared_mutex (Read-Write locks) for status lookups and converting simple flags to atomic variables. These changes result in ~10% latency reduction for UDS and significantly higher throughput under load.

* perf: implement True Zero-copy and Batching for TCP/UDS

* perf: implement True Zero-copy and Integrated Batching across all wrappers

This finalizes the architecture overhaul by integrating True Zero-copy data paths and a unified batching system for all client and server wrappers. It also adds executor access to the base Channel interface for optimized timer management.

* fix: resolve CI build errors and warnings in performance tests

This includes missing headers for system_executor, fixes conversion warnings, and addresses unused variable/result warnings in the benchmark suite.

* fix: address remaining performance test warnings and portability issues

Explicitly cast to double for performance metrics and use a static io_context for portable executor access in mock channels.

+++ test/performance/benchmark/test_packet_framer_perf.cc
@@ -53,7 +53,7 @@ TEST_F(PacketFramerPerfTest, ProcessLargeDataChunks) {
   auto end = std::chrono::high_resolution_clock::now();
   auto duration = std::chrono::duration_cast<std::chrono::microseconds>(end - start).count();
 
-  double throughput = (data.size() * 1000000.0 / duration / 1024 / 1024);
+  double throughput = (static_cast<double>(data.size()) * 1000000.0 / static_cast<double>(duration) / 1024 / 1024);
   std::cout << "Processed " << data.size() << " bytes in " << duration << " us. " << "Throughput: " << throughput
             << " MB/s. Messages: " << msg_count << std::endl;
 }
@@ -74,7 +74,7 @@ TEST_F(PacketFramerPerfTest, ProcessLargePackets) {
   auto end = std::chrono::high_resolution_clock::now();
   auto duration = std::chrono::duration_cast<std::chrono::microseconds>(end - start).count();
 
-  double throughput = (data.size() * 1000000.0 / duration / 1024 / 1024);
+  double throughput = (static_cast<double>(data.size()) * 1000000.0 / static_cast<double>(duration) / 1024 / 1024);
   std::cout << "Processed " << data.size() << " bytes in " << duration << " us. " << "Throughput: " << throughput
             << " MB/s. Messages: " << msg_count << std::endl;
 }
