+++ test/performance/benchmark/test_tcp_server_get_clients_benchmark.cc
@@ -101,5 +101,5 @@ TEST_F(TcpServerGetClientsBenchmarkTest, BenchmarkGetClients) {
 
   std::cout << "Iterations: " << iterations << std::endl;
   std::cout << "Time elapsed: " << duration_ms << " ms" << std::endl;
-  std::cout << "Ops/sec: " << (iterations * 1000.0 / duration_ms) << std::endl;
+  std::cout << "Ops/sec: " << (static_cast<double>(iterations) * 1000.0 / static_cast<double>(duration_ms)) << std::endl;
 }
